refactor(repository): share storage qualification
Move the existing journaled-filesystem qualifier and bounded kernel-option test into a shared Linux leaf. Preserve qualification policy, syscall order and public publication errors through explicit mapping. Document the module boundary without adding operation admission or changing publication behavior.

Refs #383.
